fix: utxo id output index to be u16

+++ packages/transactions/src/coders/utxo-id.ts
@@ -14,7 +14,7 @@ export class UtxoIdCoder extends StructCoder<{
   constructor() {
     super('UtxoId', {
       transactionId: new B256Coder(),
-      outputIndex: new NumberCoder('u8', { padToWordSize: true }),
+      outputIndex: new NumberCoder('u16', { padToWordSize: true }),
     });
   }
 }
